backend/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
pandas==2.0.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
scikit-learn==1.3.0
Pillow==10.0.0
transformers==4.34.0
torch==2.0.1
faiss-cpu==1.7.4.post2
numpy==1.24.4